pyproject.toml
[tool.poetry]
name = "task_manager"
version = "0.1.0"
description = ""
authors = ["Kirill Dvoretsky <shortykofficial@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
uvicorn = "^0.30.6"
django-bootstrap5 = "^24.2"
whitenoise = {extras = ["brotli"], version = "^6.7.0"}
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
coverage = "^7.6.1"
django-extensions = "^3.2.3"
ipython = "^8.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
